# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartFlow Next",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/browser": {
      "get": {
        "summary": "Info",
        "operationId": "info_api_browser_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BridgeInfo"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "browser:manage"
      }
    },
    "/api/browser/pairings": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_browser_pairings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairCode"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "browser:manage"
      }
    },
    "/api/browser/pair": {
      "post": {
        "summary": "Exchange",
        "operationId": "exchange_api_browser_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairExchange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "browser:pair"
      }
    },
    "/api/browser/agent": {
      "get": {
        "summary": "Status",
        "operationId": "status_api_browser_agent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "browser:status"
      }
    },
    "/api/browser/pairings/{pair_id}/revoke": {
      "post": {
        "summary": "Revoke",
        "operationId": "revoke_api_browser_pairings__pair_id__revoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pair Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "browser:manage"
      }
    }
  },
  "components": {
    "schemas": {
...
      "BridgeInfo": {
        "properties": {
          "extension_id": {
            "type": "string",
            "title": "Extension Id"
          },
          "extension_version": {
            "type": "string",
            "title": "Extension Version"
          },
          "helper_version": {
            "type": "string",
            "title": "Helper Version"
          },
          "host_name": {
            "type": "string",
            "title": "Host Name"
          },
          "pairings": {
            "items": {
              "$ref": "#/components/schemas/PairResponse"
            },
            "type": "array",
            "title": "Pairings"
          }
        },
        "type": "object",
        "required": [
          "extension_id",
          "extension_version",
          "helper_version",
          "host_name",
          "pairings"
        ],
        "title": "BridgeInfo"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recovery": {
            "type": "string",
            "title": "Recovery"
          },
          "http_status": {
            "type": "integer",
            "title": "Http Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "prepare",
                  "generate",
                  "save",
                  "done"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "recovery",
          "http_status",
          "trace_id",
          "stage"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InvalidField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "InvalidField": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "type"
        ],
        "title": "InvalidField"
      },
...
      "PairCode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "expires_at": {
            "type": "number",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "expires_at"
        ],
        "title": "PairCode"
      },
      "PairExchange": {
        "properties": {
          "extension_id": {
            "type": "string",
            "pattern": "^[a-p]{32}$",
            "title": "Extension Id"
          },
          "agent_token": {
            "type": "string",
            "maxLength": 100,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Agent Token"
          },
          "extension_version": {
            "type": "string",
            "title": "Extension Version"
          },
          "helper_version": {
            "type": "string",
            "title": "Helper Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "extension_id",
          "agent_token",
          "extension_version",
          "helper_version"
        ],
        "title": "PairExchange"
      },
      "PairRequest": {
        "properties": {
          "extension_id": {
            "type": "string",
            "pattern": "^[a-p]{32}$",
            "title": "Extension Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "extension_id"
        ],
        "title": "PairRequest"
      },
      "PairResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "extension_id": {
            "type": "string",
            "title": "Extension Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "last_seen": {
            "type": "number",
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "id",
          "extension_id",
          "state",
          "connected",
          "last_seen"
        ],
        "title": "PairResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}